Hoist state out of AppWithNavigationSuiteScaffold
